Split FlowHDL into the dataflow runner and the inherited FlowHDLView. The context object.
